Relocate resource.(cluster|project)
