+++ bcs-app/backend/apis/iam/resources/project.py
@@ -16,7 +16,7 @@ try:
 except Exception:
     pass
 from backend.components import ssm
-from backend.resources.project import filter_projects
+from backend.container_service.projects.base import filter_projects
 
 
 class ProjectProvider(ResourceProvider):
