+++ bcs-app/backend/bcs_web/permissions.py
@@ -22,8 +22,8 @@ from backend.apps.constants import ClusterType
 from backend.components.base import ComponentAuth
 from backend.components.iam import permissions
 from backend.components.paas_cc import PaaSCCClient
-from backend.resources.cluster.models import CtxCluster
-from backend.resources.project.models import CtxProject
+from backend.container_service.clusters.base.models import CtxCluster
+from backend.container_service.projects.base.models import CtxProject
 from backend.utils import FancyDict
 from backend.utils.cache import region
