+++ bcs-app/backend/container_service/clusters/open_apis/namespace.py
@@ -17,9 +17,9 @@ from backend.accounts import bcs_perm
 from backend.apis.views import NoAccessTokenBaseAPIViewSet
 from backend.apps.variable.models import NameSpaceVariable
 from backend.container_service.clusters.open_apis.serializers import CreateNamespaceParamsSLZ
+from backend.container_service.projects.base.constants import ProjectKind
 from backend.resources.namespace import utils as ns_utils
 from backend.resources.namespace.constants import K8S_PLAT_NAMESPACE, K8S_SYS_NAMESPACE
-from backend.resources.project.constants import ProjectKind
 from backend.utils.error_codes import error_codes
