+++ bcs-app/backend/container_service/clusters/open_apis/node.py
@@ -14,8 +14,8 @@
 from rest_framework.response import Response
 
 from backend.apis.views import NoAccessTokenBaseAPIViewSet
+from backend.container_service.clusters.base.utils import set_mesos_node_labels
 from backend.container_service.clusters.open_apis.serializers import CreateNodeLabelsSLZ
-from backend.resources.cluster.utils import set_mesos_node_labels
 
 
 class NodeLabelsViewSet(NoAccessTokenBaseAPIViewSet):
