+++ bcs-app/backend/container_service/clusters/views/node.py
@@ -29,13 +29,13 @@ from backend.components.bcs import k8s, mesos
 from backend.container_service.clusters import constants as cluster_constants
 from backend.container_service.clusters import serializers as node_serializers
 from backend.container_service.clusters import utils as cluster_utils
+from backend.container_service.clusters.base.utils import get_cluster_nodes
 from backend.container_service.clusters.constants import DEFAULT_SYSTEM_LABEL_KEYS
 from backend.container_service.clusters.driver import BaseDriver
 from backend.container_service.clusters.models import CommonStatus, NodeLabel, NodeStatus, NodeUpdateLog
 from backend.container_service.clusters.module_apis import get_cluster_node_mod, get_cmdb_mod, get_gse_mod
 from backend.container_service.clusters.serializers import NodeLabelParamsSLZ
 from backend.container_service.clusters.utils import cluster_env_transfer, custom_paginator, status_transfer
-from backend.resources.cluster.utils import get_cluster_nodes
 from backend.utils.errcodes import ErrorCode
 from backend.utils.error_codes import error_codes
 from backend.utils.funutils import convert_mappings
