+++ bcs-app/backend/container_service/core/ctx_models.py
@@ -11,7 +11,7 @@
 # an "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the License for the
 # specific language governing permissions and limitations under the License.
 #
-"""保存 resource 模型相关内容"""
+"""保存上下文模型相关内容"""
 from backend.components.base import ComponentAuth
 from backend.components.collection import ComponentCollection
