+++ bcs-app/backend/resources/custom_object/custom_object.py
@@ -16,7 +16,7 @@ from typing import Optional
 from django.utils.translation import ugettext_lazy as _
 from kubernetes.dynamic.resource import ResourceInstance
 
-from backend.resources.cluster.models import CtxCluster
+from backend.container_service.clusters.base.models import CtxCluster
 from backend.utils.error_codes import error_codes
 
 from ..resource import ResourceClient
