+++ bcs-app/backend/tests/resources/hpa/test_hpa.py
@@ -17,7 +17,7 @@ import pytest
 import yaml
 from kubernetes.dynamic.exceptions import ResourceNotFoundError
 
-from backend.resources.cluster.models import CtxCluster
+from backend.container_service.clusters.base.models import CtxCluster
 from backend.resources.hpa import hpa as hpa_client
 from backend.utils.basic import getitems
